refactor: rename forbidden_usernames module to usernames
